+++ src/components/AppIndex.vue
@@ -1,26 +1,47 @@
 <template>
     <div>
         <div class="kw">
-            <transition-group v-if="keywords.length" class="keywords" name="keyword-list" tag="ul">
-                <RecipeKeyword v-for="keyword in keywords"
-                    :key="keyword.name"
-                    :name="keyword.name"
-                    :count="keyword.count"
-                    :title="keywordContainedInVisibleRecipes(keyword) ? t('cookbook','Toggle keyword') : t('cookbook','Keyword not contained in visible recipes')"
-                    v-on:keyword-clicked="keywordClicked(keyword)"
-                    :class="{keyword, active : keywordFilter.includes(keyword.name), disabled : !keywordContainedInVisibleRecipes(keyword)}"
+            <transition-group v-if="uniqKeywords.length > 0" class="keywords" name="keyword-list" tag="ul">
+                <RecipeKeyword v-for="keywordObj in selectedKeywords"
+                    :key="keywordObj.name"
+                    :name="keywordObj.name"
+                    :count="keywordObj.count"
+                    :title="t('cookbook','Toggle keyword')"
+                    @keyword-clicked="keywordClicked(keywordObj)"
+                    class="keyword active"
                     />
+                <RecipeKeyword v-for="keywordObj in selectableKeywords"
+                    :key="keywordObj.name"
+                    :name="keywordObj.name"
+                    :count="keywordObj.count"
+                    :title="t('cookbook','Toggle keyword')"
+                    @keyword-clicked="keywordClicked(keywordObj)"
+                    class="keyword"
+                    />
+                <RecipeKeyword v-for="keywordObj in unavailableKeywords"
+                    :key="keywordObj.name"
+                    :name="keywordObj.name"
+                    :count="keywordObj.count"
+                    :title="t('cookbook','Keyword not contained in visible recipes')"
+                    @keyword-clicked="keywordClicked(keywordObj)"
+                    class="keyword disabled"
+                    />
+                
             </transition-group>
         </div>
         <ul class="recipes">
-            <li v-for="(recipe, index) in filteredRecipes" :key="recipe.recipe_id" v-show="recipeVisible(index)">
-                <router-link :to="'/recipe/'+recipe.recipe_id">
-                     <lazy-picture v-if="recipe.imageUrl"
+            <li
+                v-for="(recipeObj, index) in recipeObjects" 
+                :key="recipeObj.recipe.recipe_id"
+                v-show="recipeObj.show"
+                >
+                <router-link :to="'/recipe/' + recipeObj.recipe.recipe_id">
+                     <lazy-picture v-if="recipeObj.recipe.imageUrl"
                         class="recipe-thumbnail"
-                        :lazy-src="recipe.imageUrl"
-                        :blurred-preview-src="recipe.imagePlaceholderUrl"
+                        :lazy-src="recipeObj.recipe.imageUrl"
+                        :blurred-preview-src="recipeObj.recipe.imagePlaceholderUrl"
                         :width="105" :height="105"/>
-                    <span>{{ recipe.name }}</span>
+                    <span>{{ recipeObj.recipe.name }}</span>
                 </router-link>
             </li>
         </ul>
@@ -39,34 +60,125 @@ export default {
     },
     data () {
         return {
+            // String-based filters applied to the list
             filters: "",
+            // The known recipes in the cookbook
             recipes: [],
-            keywords: [],
+            // All keyword to filter the recipes for (conjunctively)
             keywordFilter: [],
         }
     },
     computed: {
-        filteredRecipes () {
-            if (!this.filters) {
-                return this.recipes
-            } else if (this.recipes.length) {
-                let filtered = []
-                for (let i=0; i<this.recipes.length; i++) {
-                    if (this.recipes[i].name.toLowerCase().indexOf(this.filters.toLowerCase()) >= 0) {
-                        filtered.push(this.recipes[i])
-                    }
+        // An array of all keywords in all recipes. These are neither sorted nor unique
+        rawKeywords() {
+            var keywordArray = this.recipes.map(function(r){
+                if(! 'keywords' in r) {
+                    return []
                 }
-                return filtered
-            }
+                if(r.keywords != null){
+                    return r.keywords.split(',')
+                } else {
+                    return []
+                }
+            })
+            return [].concat(... keywordArray)
         },
-    },
-    watch: {
-        keywordFilter: {
-            handler: function() {
-                this.sortKeywords()
-            },
-            deep: true
-        }
+        // An array of sorted and unique keywords over all the recipes
+        uniqKeywords() {
+            function uniqFilter(value, index, self) {
+                return self.indexOf(value) === index
+            }
+            return this.rawKeywords.sort().filter(uniqFilter)
+        },
+        // An array of objects that contain the keywords plus a count of recipes asociated with these keywords
+        keywordsWithCount() {
+            let $this = this
+            return this.uniqKeywords.map(function (kw){
+                return {
+                    'name': kw,
+                    'count': $this.rawKeywords.filter((kw2) => kw === kw2).length
+                }
+            }).sort(function (k1, k2) {
+                if(k1.count != k2.count) {
+                    // Distinguish by number
+                    return k2.count - k1.count
+                }
+                // Distinguish by keyword name
+                return (k1.name.toLowerCase() > k2.name.toLowerCase()) ? 1 : -1
+            })
+        },
+        // An array of keyword objects that are currently in use for filtering
+        selectedKeywords() {
+            return this.keywordsWithCount.filter((kw) => this.keywordFilter.includes(kw.name))
+        },
+        // An array of those keyword objects that are currently not in use for filtering
+        unselectedKeywords() {
+            return this.keywordsWithCount.filter((kw) => ! this.selectedKeywords.includes(kw))
+        },
+        // An array of all recipes that are part in all filtered keywords
+        recipesFilteredByKeywords() {
+            let $this = this
+            return this.recipes.filter(function (r) {
+                if($this.keywordFilter.length == 0) {
+                    // No filtering, keep all
+                    return true
+                }
+                
+                if(r.keywords === null){
+                    return false
+                }
+                
+                function keywordInRecipePresent(kw,r) {
+                    if(! r.keywords) {
+                        return false;
+                    }
+                    
+                    let keywords = r.keywords.split(',')
+                    return keywords.includes(kw.name)
+                }
+                
+                return $this.selectedKeywords.map((kw) => keywordInRecipePresent(kw, r)).reduce((l,r) => l && r)
+            })
+        },
+        // An array of the finally filtered recipes, that is both filtered for keywords as well as string-based name filtering
+        filteredRecipes () {
+            let ret = this.recipesFilteredByKeywords
+            let $this = this
+            
+            if(this.filters){
+                ret = ret.filter(function (r) {
+                    return r.name.toLowerCase().includes($this.filters.toLowerCase())
+                })
+            }
+            
+            return ret
+        },
+        // An array of keywords that are yet unselected but some visible recipes are associated
+        selectableKeywords() {
+            if (this.unselectedKeywords.length === 0) {
+                return []
+            }
+            
+            let $this = this
+            return this.unselectedKeywords.filter(function (kw) {
+                return $this.filteredRecipes.map(function (r) {
+                    return r.keywords && r.keywords.split(',').includes(kw.name)
+                }).reduce((l,r) => l || r, false)
+            })
+        },
+        // An array of known keywords that are not associated with any visible recipe
+        unavailableKeywords() {
+            return this.unselectedKeywords.filter((kw) => ! this.selectableKeywords.includes(kw))
+        },
+        // An array of recipe objects of all recipes with links to the recipes and a property if the recipe is to be shown 
+        recipeObjects() {
+            return this.recipes.map(function (r) {
+                return {
+                    'recipe': r,
+                    'show': this.filteredRecipes.includes(r)
+                }
+            }, this)
+        },
     },
     methods: {
         /**
@@ -81,19 +193,6 @@ export default {
             }
         },
         /**
-         * Check if a keyword exists in the currently visible recipes.
-         */
-        keywordContainedInVisibleRecipes: function(keyword) {
-            for (let i=0; i<this.recipes.length; ++i) {
-                if (this.recipeVisible(i)
-                    && this.recipes[i].keywords
-                    && this.recipes[i].keywords.split(',').includes(keyword.name)) {
-                    return true
-                }
-            }
-            return false
-        },
-        /**
          * Load all recipes from the database
          */
         loadAll: function () {
@@ -101,7 +200,6 @@ export default {
             var $this = this
             $.get(this.$window.baseUrl + '/api/recipes').done(function (recipes) {
                 $this.recipes = recipes
-                $this.setKeywords(recipes)
                 deferred.resolve()
                 // Always set page name last
                 $this.$store.dispatch('setPage', { page: 'index' })
@@ -111,64 +209,6 @@ export default {
                 $this.$store.dispatch('setPage', { page: 'index' })
             })
             return deferred.promise()
-        },
-        /**
-         * Check if recipe should be displayed, depending on selected keyword filter.
-         * Returns true if recipe contains all selected keywords.
-         */
-        recipeVisible: function(index) {
-            if (this.keywordFilter.length == 0) {
-                return true
-            } else {
-                if (!this.recipes[index].keywords) {
-                    return false
-                }
-                let kw_array = this.recipes[index].keywords.split(',')
-                return this.keywordFilter.every(kw => kw_array.includes(kw))
-            }
-        },
-        /**
-         * Extract and set list of keywords from the returned recipes.
-         */
-        setKeywords: function(recipes) {
-            this.keywords = []
-            if ((recipes.length) > 0) {
-                recipes.forEach(recipe => {
-                    if(recipe['keywords']) {
-                        recipe['keywords'].split(',').forEach(kw => {
-                            const idx = this.keywords.findIndex(el => el.name == kw)
-                            if (idx > -1) {
-                                this.keywords[idx].count++
-                            } else {
-                                this.keywords.push({name: kw, count: 1})
-                            }
-                        })
-                    }
-                })
-                this.sortKeywords()
-            }
-        },
-        /**
-         * Sort keywords.
-         */
-        sortKeywords: function() {
-            // Sort by number of recipes containing keyword
-            this.keywords = this.keywords.sort((k1, k2) => k2.count - k1.count)
-
-            // Move selected keywords to the front and unselectable to the end
-            let selected_kw = [], selectable_kw = [], unavailable_kw = []
-            this.keywords.forEach(kw => {
-                if (this.keywordFilter.includes(kw.name)) {
-                    selected_kw.push(kw)
-                }
-                else if (this.keywordContainedInVisibleRecipes(kw)) {
-                    selectable_kw.push(kw)
-                }
-                else {
-                    unavailable_kw.push(kw)
-                }
-            })
-            this.keywords = selected_kw.concat(selectable_kw.concat(unavailable_kw))
         },
     },
     mounted () {